requirements.txt
# These requirements are only necessary when developing on Oscar.

# development
Werkzeug==0.12.2
django-debug-toolbar==1.8
django-extensions==1.8.1
psycopg2==2.7.3

# Sandbox
Pillow==4.2.1
Whoosh==2.7.4
django-environ==0.4.3
django-redis==4.8.0
pysolr==3.6.0
redis==2.10.5
requests==2.18.3
uWSGI==2.0.15
whitenoise==3.3.0

# Linting
flake8==3.4.1
flake8-debugger==1.4.0
flake8-blind-except==0.1.1
isort==4.2.15

# Helpers
pyprof2calltree==1.4.3
ipdb==0.10.3
ipython==5.3.0
transifex-client==0.12.4

# Country data
pycountry==17.5.14

xapian-haystack==2.1.1
django-compressor==2.1
jieba==0.39